requirements.txt
python-dotenv
openai
azure-search-documents
azure-identity
azure-core
